+++ src/features/surveys/components/AnswerTableRow/AnswerTableRow.tsx
