+++ src/features/surveys/components/EmojiButton/EmojiButton.tsx
