+++ src/features/surveys/interfaces/AnswerData.ts
@@ -1,0 +1,6 @@
+export interface AnswerData {
+  id: string;
+  answerDate: string;
+  selectedIcon: string;
+  answer: string;
+}
