+++ src/pages/_app.tsx
@@ -4,18 +4,24 @@ import { AppProps } from 'next/app';
 import { Toaster } from 'react-hot-toast';
 import PageLayout from '../layout/PageLayout';
 import Head from 'next/head';
+import { ApplicationContext } from 'src/features/application/context';
+import { useApplicationManager } from 'src/features/application/manager';
 
 function MyApp({ Component, pageProps }: AppProps) {
+  const manager = useApplicationManager();
+
   return (
     <>
       <Head>
         <title>Employee Pulse</title>
       </Head>
-      <PageLayout>
-        <Toaster position="bottom-center" />
+      <ApplicationContext.Provider value={manager}>
+        <PageLayout>
+          <Toaster position="bottom-center" />
 
-        <Component {...pageProps} />
-      </PageLayout>
+          <Component {...pageProps} />
+        </PageLayout>
+      </ApplicationContext.Provider>
     </>
   );
 }
