+++ src/pages/survey/[surveyId]/index.tsx
@@ -1,99 +1,25 @@
-import { addDoc, collection, doc, getDoc } from 'firebase/firestore';
-import { useEffect, useState } from 'react';
-import toast from 'react-hot-toast';
-
-import { useRouter } from 'next/router';
-
-import { db } from '../../../firebase';
 import Head from 'next/head';
 import Button, {
   ButtonVariant,
   ButtonSize,
-} from 'src/components/Button/Button';
-import EmojiButton from 'src/components/EmojiButton/EmojiButton';
-import Header from 'src/components/Header/Header';
-import Loader from 'src/components/Loader/Loader';
+} from 'src/shared/components/Button/Button';
+import EmojiButton from 'src/features/surveys/components/EmojiButton/EmojiButton';
+import Header from 'src/shared/components/Header/Header';
+import Loader from 'src/shared/components/Loader/Loader';
+import { useSurveyAnswerManager } from 'src/features/surveys/managers/surveyAnswerManager';
 
-function SurveyPage() {
-  const router = useRouter();
-
-  const { surveyId } = router.query as { surveyId: string };
-
-  const [question, setQuestion] = useState('');
-  const [answer, setAnswer] = useState('');
-  const [icons, setIcons] = useState<string[]>([]);
-  const [selectedIcon, setSelectedIcon] = useState<string | null>('');
-  const [buttonDisable, setButtonDisable] = useState(false);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    if (surveyId) {
-      getSurveyData();
-    }
-  }, [surveyId]);
-
-  const getSurveyData = async () => {
-    const surveyData = await getDoc(doc(db, 'surveys', surveyId!));
-    if (!surveyData.exists()) {
-      router.replace('/');
-      return;
-    }
-
-    if (
-      surveyData.data()?.startDate.toDate().toISOString() >
-      new Date().toISOString()
-    ) {
-      toast.error('Survey is not opened yet');
-      router.replace('/');
-      return;
-    }
-
-    if (
-      surveyData.data()?.endDate.toDate().toISOString() <
-      new Date().toISOString()
-    ) {
-      toast.error('Survey is closed');
-      router.replace('/');
-      return;
-    }
-
-    setQuestion(surveyData.data()?.title);
-    setIcons(surveyData.data()?.pack);
-    setIsLoading(false);
-  };
-
-  const handleIconClick = (e: React.MouseEvent<HTMLButtonElement>) => {
-    const target = e.target as HTMLElement;
-    setSelectedIcon(target.textContent);
-  };
-
-  const handleSave = async () => {
-    setButtonDisable(true);
-    setIsLoading(true);
-
-    try {
-      if (!surveyId) {
-        toast.error('Survey ID not found');
-        throw new Error('Survey ID not found');
-      }
-      await addDoc(collection(db, 'surveys', surveyId, 'answers'), {
-        selectedIcon,
-        answer,
-        answerDate: new Date(),
-      });
-      toast.success('The reply has been sent');
-      router.replace('/');
-    } catch (error) {
-      toast.error('Error occured');
-    } finally {
-      setButtonDisable(false);
-      setIsLoading(false);
-    }
-  };
-
-  const handleInputAnswer = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setAnswer(e.target.value);
-  };
+function AnswerPage() {
+  const {
+    isLoading,
+    question,
+    icons,
+    selectedIcon,
+    handleIconClick,
+    answer,
+    handleInputAnswer,
+    buttonDisable,
+    handleSave,
+  } = useSurveyAnswerManager();
 
   return (
     <>
@@ -138,4 +64,4 @@ function SurveyPage() {
   );
 }
 
-export default SurveyPage;
+export default AnswerPage;
