+++ src/shared/hooks/useCloseComponent.ts
